lift dog image fixtures to integration test folder
